Move setup to /setup and tighten validation
