+++ tinyoffice/package.json
@@ -1,7 +1,6 @@
 {
-  "name": "mission-control",
+  "name": "tinyoffice",
   "version": "0.1.0",
-  "private": true,
   "scripts": {
     "dev": "next dev --webpack",
     "build": "next build",
