+++ tinyoffice/src/components/app-shell.tsx
@@ -1,0 +1,18 @@
+"use client";
+
+import { usePathname } from "next/navigation";
+import { Sidebar } from "@/components/sidebar";
+
+export function AppShell({ children }: { children: React.ReactNode }) {
+  const pathname = usePathname();
+  const hideSidebar = pathname === "/setup";
+
+  return (
+    <div className="flex h-screen overflow-hidden">
+      {!hideSidebar && <Sidebar />}
+      <main className="flex-1 overflow-y-auto">
+        {children}
+      </main>
+    </div>
+  );
+}
